old/my-web-app/requirements.txt
Flask
python-dotenv
pyotp
logzero
requests
json
SmartApi